refactor: split widgets.py into widgets/ package (_base, _widgets2d, _widgets1d)
