+++ anyplotlib/widgets/__init__.py
@@ -1,0 +1,15 @@
+from anyplotlib.widgets._base import Widget
+from anyplotlib.widgets._widgets2d import (
+    RectangleWidget, CircleWidget, AnnularWidget,
+    CrosshairWidget, PolygonWidget, LabelWidget,
+)
+from anyplotlib.widgets._widgets1d import (
+    VLineWidget, HLineWidget, RangeWidget, PointWidget,
+)
+
+__all__ = [
+    "Widget",
+    "RectangleWidget", "CircleWidget", "AnnularWidget",
+    "CrosshairWidget", "PolygonWidget", "LabelWidget",
+    "VLineWidget", "HLineWidget", "RangeWidget", "PointWidget",
+]
